+++ packages/xml-views-completion/test/unit/utils.ts
@@ -12,6 +12,7 @@ import { getSuggestions, SuggestionProviders } from "@xml-tools/content-assist";
 
 import { UI5XMLViewCompletion } from "../../api";
 import { Context } from "@ui5-language-assistant/context";
+import { DEFAULT_UI5_FRAMEWORK } from "@ui5-language-assistant/constant";
 
 export function testSuggestionsScenario(opts: {
   xmlText: string;
@@ -90,7 +91,7 @@ export const getDefaultContext = (ui5Model: UI5SemanticModel): Context => {
     },
     services: {},
     yamlDetails: {
-      framework: "SAPUI5",
+      framework: DEFAULT_UI5_FRAMEWORK,
       version: undefined,
     },
     viewFiles: {},