+++ packages/fe/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 4.0.30
+
+### Patch Changes
+
+- Updated dependencies [b2a8d01]
+  - @ui5-language-assistant/xml-views-completion@4.0.27
+  - @ui5-language-assistant/xml-views-validation@4.0.21
+  - @ui5-language-assistant/logic-utils@4.0.19
+  - @ui5-language-assistant/context@4.0.27
+
 ## 4.0.29
 
 ### Patch Changes